Re-organize conversation integration
