+++ tests/components/conversation/__init__.py
@@ -5,11 +5,16 @@ from __future__ import annotations
 from typing import Literal
 
 from homeassistant.components import conversation
+from homeassistant.components.conversation.agent import (
+    ConversationInput,
+    ConversationResult,
+)
 from homeassistant.components.homeassistant.exposed_entities import (
     DATA_EXPOSED_ENTITIES,
     ExposedEntities,
     async_expose_entity,
 )
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers import intent
 
 
@@ -30,24 +35,22 @@ class MockAgent(conversation.AbstractConversationAgent):
         """Return a list of supported languages."""
         return self._supported_languages
 
-    async def async_process(
-        self, user_input: conversation.ConversationInput
-    ) -> conversation.ConversationResult:
+    async def async_process(self, user_input: ConversationInput) -> ConversationResult:
         """Process some text."""
         self.calls.append(user_input)
         response = intent.IntentResponse(language=user_input.language)
         response.async_set_speech(self.response)
-        return conversation.ConversationResult(
+        return ConversationResult(
             response=response, conversation_id=user_input.conversation_id
         )
 
 
-def expose_new(hass, expose_new):
+def expose_new(hass: HomeAssistant, expose_new: bool):
     """Enable exposing new entities to the default agent."""
     exposed_entities: ExposedEntities = hass.data[DATA_EXPOSED_ENTITIES]
     exposed_entities.async_set_expose_new_entities(conversation.DOMAIN, expose_new)
 
 
-def expose_entity(hass, entity_id, should_expose):
+def expose_entity(hass: HomeAssistant, entity_id: str, should_expose: bool):
     """Expose an entity to the default agent."""
     async_expose_entity(hass, conversation.DOMAIN, entity_id, should_expose)
