+++ tests/components/conversation/test_trigger.py
@@ -6,6 +6,8 @@ import pytest
 import voluptuous as vol
 
 from homeassistant.components import conversation
+from homeassistant.components.conversation import default_agent
+from homeassistant.components.conversation.agent import ConversationInput
 from homeassistant.core import Context, HomeAssistant
 from homeassistant.helpers import trigger
 from homeassistant.setup import async_setup_component
@@ -514,11 +516,11 @@ async def test_trigger_with_device_id(hass: HomeAssistant) -> None:
         },
     )
 
-    agent = await conversation._get_agent_manager(hass).async_get_agent()
-    assert isinstance(agent, conversation.DefaultAgent)
+    agent = await conversation.get_agent_manager(hass).async_get_agent()
+    assert isinstance(agent, default_agent.DefaultAgent)
 
     result = await agent.async_process(
-        conversation.ConversationInput(
+        ConversationInput(
             text="test sentence",
             context=Context(),
             conversation_id=None,
